# docs/swagger.yaml
basePath: /api/v1/
definitions:
  lib.Page:
    properties:
      first:
        description: indicate first value
        example: true
        type: boolean
      items:
        type: object
      last:
        description: indicate last value
        example: false
        type: boolean
      max_page:
        description: maximum pages for current schema
        example: 9
        type: integer
      page:
        description: current page, start from zero
        example: 0
        type: integer
      size:
        description: size per page, default `10`
        example: 10
        type: integer
      total:
        description: total data without filters
        example: 100
        type: integer
      total_pages:
        description: total pages
        example: 10
        type: integer
      visible:
        description: current length
        example: 10
        type: integer
    type: object
  lib.Response:
    properties:
      message:
        description: response message
        type: string
      status:
        description: http status
        type: integer
    type: object
  model.Asset:
    properties:
      absolute_path:
        description: AbsolutePath
        example: image.png
        type: string
      creator_id:
        description: creator id
        format: uuid
        type: string
      description:
        description: Description
        example: string
        type: string
      file_path:
        description: File Path
        type: string
      file_size:
        description: File Size
        format: float
        type: number
      filename:
        description: Filename
        example: image.png
        type: string
      modifier_id:
        description: modifier id
        format: uuid
        type: string
      original_filename:
        description: OriginalFilename
        example: image.png
        type: string
      relative_path:
        description: RelativePath
        example: image.png
        type: string
      title:
        description: Title
        example: Image
        type: string
    type: object
  model.Brand:
    properties:
      brand_code:
        example: XM
        type: string
      brand_name:
        example: Iphone
        type: string
    required:
    - brand_code
    - brand_name
    type: object
...
  model.Cart:
    properties:
      customer:
        $ref: '#/definitions/model.Customer'
      customer_id:
        format: uuid
        type: string
      product:
        $ref: '#/definitions/model.Product'
      product_id:
        format: uuid
        type: string
      qty:
        type: integer
    type: object
  model.CartRequest:
    properties:
      customer_id:
        format: uuid
        type: string
      product_id:
        format: uuid
        type: string
      qty:
        type: integer
    type: object
  model.CartUpdate:
    properties:
      qty:
        type: integer
    type: object
  model.Category:
    properties:
      category_code:
        example: HP
        type: string
      category_name:
        example: Handphone
        type: string
    required:
    - category_code
    - category_name
    type: object
  model.CategoryAPI:
    properties:
      category_code:
        example: HP
        type: string
      category_name:
        example: Handphone
        type: string
    required:
    - category_code
    - category_name
    type: object
  model.Customer:
    properties:
      address:
        description: Address
        example: Jl. Aria Putra No.88, RT.09/RW.01, Sawah Baru, Kec. Ciputat, Kota
          Tangerang Selatan, Banten 15414
        type: string
      alternate_number:
        description: Alternate Number
        example: "08123456789"
        type: string
      city_id:
        description: City ID
        format: uuid
        type: string
      creator_id:
        description: creator id
        format: uuid
        type: string
      customer_name:
        description: customer Name                                                                                //
          Type
        example: Walk-in-customers
        type: string
      email:
        description: Email
        example: walk-in-customer@mail.com
        type: string
      mobile:
        description: Mobile
        example: "08123456789"
        type: string
      modifier_id:
        description: modifier id
        format: uuid
        type: string
      province_id:
        description: Province ID
        format: uuid
        type: string
      subdistrict_id:
        description: Subdistrict ID
        format: uuid
        type: string
      user:
        $ref: '#/definitions/model.User'
      user_id:
        format: uuid
        type: string
      website:
        description: Website
        example: www.walk-in-customer.com
        type: string
      zip_code:
        description: Zip Code
        example: "15414"
        type: string
    required:
    - customer_name
    type: object
  model.CustomerAPI:
    properties:
      address:
        description: Address
        example: Jl. Aria Putra No.88, RT.09/RW.01, Sawah Baru, Kec. Ciputat, Kota
          Tangerang Selatan, Banten 15414
        type: string
      alternate_number:
        description: Alternate Number
        example: "08123456789"
        type: string
      city_id:
        description: City ID
        format: uuid
        type: string
      customer_name:
        description: customer Name                                                                                //
          Type
        example: Walk-in-customers
        type: string
      email:
        description: Email
        example: walk-in-customer@mail.com
        type: string
      mobile:
        description: Mobile
        example: "08123456789"
        type: string
      province_id:
        description: Province ID
        format: uuid
        type: string
      subdistrict_id:
        description: Subdistrict ID
        format: uuid
        type: string
      user_id:
        format: uuid
        type: string
      website:
        description: Website
        example: www.walk-in-customer.com
        type: string
      zip_code:
        description: Zip Code
        example: "15414"
        type: string
    required:
    - customer_name
    type: object
...
  model.Product:
    properties:
      brand:
        $ref: '#/definitions/model.Brand'
      brand_id:
        format: uuid
        type: string
      category:
        $ref: '#/definitions/model.Category'
      category_id:
        format: uuid
        type: string
      description:
        example: Samsung Desc
        type: string
      price:
        example: 10000
        type: number
      product_asset:
        $ref: '#/definitions/model.ProductAsset'
      product_name:
        example: Samsung
        type: string
      quantity:
        example: 10
        type: integer
      sku:
        example: PR6900000007
        type: string
    required:
    - description
    - product_name
    type: object
...
  model.ProductAsset:
    properties:
      file_name:
        type: string
      file_path:
        type: string
      is_primary:
        type: boolean
      product_id:
        format: uuid
        type: string
    type: object
...
  model.User:
    properties:
      activated_at:
        format: date-time
        type: string
      asset:
        $ref: '#/definitions/model.Asset'
      asset_id:
        type: string
      creator_id:
        description: creator id
        format: uuid
        type: string
      email:
        type: string
      fullname:
        type: string
      is_activated:
        type: boolean
      is_owner:
        type: boolean
      is_password_system_generated:
        type: boolean
      last_login:
        format: date-time
        type: string
      mobile:
        type: string
      modifier_id:
        description: modifier id
        format: uuid
        type: string
      otp_enabled:
        type: boolean
      otp_expiration:
        format: date-time
        type: string
      password_expiration:
        format: date-time
        type: string
      password_last_change:
        format: date-time
        type: string
      reset_password_expiration:
        format: date-time
        type: string
      username:
        type: string
      verification_expiration:
        format: date-time
        type: string
    type: object
...
info:
  contact:
    email: erfanpratama321@gmail.com
    name: Irfan Pratama
  description: API Documentation
  termsOfService: https://dospecs.monstercode.net/en/guide/tnc.html
  title: User Services
  version: 1.0.0
paths:
...
  /Customers/{id}:
    delete:
      consumes:
      - application/json
      description: Delete Customer by id
      parameters:
      - description: Customer ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/lib.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/lib.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/lib.Response'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/lib.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/lib.Response'
        default:
          description: ""
          schema:
            $ref: '#/definitions/lib.Response'
      security:
      - ApiKeyAuth: []
      summary: Delete Customer by id
      tags:
      - Customer
...
  /carts:
    get:
      consumes:
      - application/json
      description: List of Cart
      parameters:
      - description: Page number start from zero
        in: query
        name: page
        type: integer
      - description: Size per page, default `0`
        in: query
        name: size
        type: integer
      - description: Sort by field, adding dash (`-`) at the beginning means descending
          and vice versa
        in: query
        name: sort
        type: string
      - description: Select specific fields with comma separated
        in: query
        name: fields
        type: string
      - description: custom filters, see [more details](https://github.com/morkid/paginate#filter-format)
        in: query
        name: filters
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: List of Cart
          schema:
            allOf:
            - $ref: '#/definitions/lib.Page'
            - properties:
                items:
                  items:
                    $ref: '#/definitions/model.Cart'
                  type: array
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/lib.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/lib.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/lib.Response'
        default:
          description: ""
          schema:
            $ref: '#/definitions/lib.Response'
      security:
      - ApiKeyAuth: []
      summary: List of Cart
      tags:
      - Cart
    post:
      consumes:
      - application/json
      description: Create new Cart
      parameters:
      - description: Cart data
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/model.CartRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Cart data
          schema:
            $ref: '#/definitions/model.Cart'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/lib.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/lib.Response'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/lib.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/lib.Response'
        default:
          description: ""
          schema:
            $ref: '#/definitions/lib.Response'
      security:
      - ApiKeyAuth: []
      summary: Create new Cart
      tags:
      - Cart
  /carts/{id}:
    delete:
      consumes:
      - application/json
      description: Delete Cart by id
      parameters:
      - description: Cart ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/lib.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/lib.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/lib.Response'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/lib.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/lib.Response'
        default:
          description: ""
          schema:
            $ref: '#/definitions/lib.Response'
      security:
      - ApiKeyAuth: []
      summary: Delete Cart by id
      tags:
      - Cart
    put:
      consumes:
      - application/json
      description: Update Cart by id
      parameters:
      - description: Cart ID
        in: path
        name: id
        required: true
        type: string
      - description: Cart data
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/model.CartUpdate'
      produces:
      - application/json
      responses:
        "200":
          description: Cart data
          schema:
            $ref: '#/definitions/model.Cart'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/lib.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/lib.Response'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/lib.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/lib.Response'
        default:
          description: ""
          schema:
            $ref: '#/definitions/lib.Response'
      security:
      - ApiKeyAuth: []
      summary: Update Cart by id
      tags:
      - Cart
  /categories:
    get:
      consumes:
      - application/json
      description: List of Category
      parameters:
      - description: Page number start from zero
        in: query
        name: page
        type: integer
      - description: Size per page, default `0`
        in: query
        name: size
        type: integer
      - description: Sort by field, adding dash (`-`) at the beginning means descending
          and vice versa
        in: query
        name: sort
        type: string
      - description: Select specific fields with comma separated
        in: query
        name: fields
        type: string
      - description: custom filters, see [more details](https://github.com/morkid/paginate#filter-format)
        in: query
        name: filters
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: List of Category
          schema:
            allOf:
            - $ref: '#/definitions/lib.Page'
            - properties:
                items:
                  items:
                    $ref: '#/definitions/model.Category'
                  type: array
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/lib.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/lib.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/lib.Response'
        default:
          description: ""
          schema:
            $ref: '#/definitions/lib.Response'
      security:
      - ApiKeyAuth: []
      summary: List of Category
      tags:
      - Category
    post:
      consumes:
      - application/json
      description: Create new Category
      parameters:
      - description: Category data
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/model.CategoryAPI'
      produces:
      - application/json
      responses:
        "201":
          description: Category data
          schema:
            $ref: '#/definitions/model.Category'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/lib.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/lib.Response'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/lib.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/lib.Response'
        default:
          description: ""
          schema:
            $ref: '#/definitions/lib.Response'
      security:
      - ApiKeyAuth: []
      summary: Create new Category
      tags:
      - Category
  /categories/{id}:
    delete:
      consumes:
      - application/json
      description: Delete Category by id
      parameters:
      - description: Category ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/lib.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/lib.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/lib.Response'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/lib.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/lib.Response'
        default:
          description: ""
          schema:
            $ref: '#/definitions/lib.Response'
      security:
      - ApiKeyAuth: []
      summary: Delete Category by id
      tags:
      - Category
    get:
      consumes:
      - application/json
      description: Get a Category by id
      parameters:
      - description: Category ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Category data
          schema:
            $ref: '#/definitions/model.Category'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/lib.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/lib.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/lib.Response'
        default:
          description: ""
          schema:
            $ref: '#/definitions/lib.Response'
      security:
      - ApiKeyAuth: []
      summary: Get a Category by id
      tags:
      - Category
    put:
      consumes:
      - application/json
      description: Update Category by id
      parameters:
      - description: Category ID
        in: path
        name: id
        required: true
        type: string
      - description: Category data
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/model.CategoryAPI'
      produces:
      - application/json
      responses:
        "200":
          description: Category data
          schema:
            $ref: '#/definitions/model.Category'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/lib.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/lib.Response'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/lib.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/lib.Response'
        default:
          description: ""
          schema:
            $ref: '#/definitions/lib.Response'
      security:
      - ApiKeyAuth: []
      summary: Update Category by id
      tags:
      - Category
...
  /customers/{id}:
    get:
      consumes:
      - application/json
      description: Get a Customer by id
      parameters:
      - description: Customer ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Customer data
          schema:
            $ref: '#/definitions/model.Customer'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/lib.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/lib.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/lib.Response'
        default:
          description: ""
          schema:
            $ref: '#/definitions/lib.Response'
      security:
      - ApiKeyAuth: []
      summary: Get a Customer by id
      tags:
      - Customer
    put:
      consumes:
      - application/json
      description: Update Customer by id
      parameters:
      - description: Customer ID
        in: path
        name: id
        required: true
        type: string
      - description: Customer data
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/model.CustomerAPI'
      produces:
      - application/json
      responses:
        "200":
          description: Customer data
          schema:
            $ref: '#/definitions/model.Customer'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/lib.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/lib.Response'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/lib.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/lib.Response'
        default:
          description: ""
          schema:
            $ref: '#/definitions/lib.Response'
      security:
      - ApiKeyAuth: []
      summary: Update Customer by id
      tags:
      - Customer
...
schemes:
- https
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
  TokenKey:
    in: header
    name: x-Token
    type: apiKey
swagger: "2.0"